Move Place Details -> components
